+++ Controllers/todolist.controller.js
@@ -1,5 +1,5 @@
-const TodolistModel = require("../Models/todolist.model");
-const UserModel = require("../Models/user.model");
+const TodolistModel = require('../Models/todolist.model')
+const UserModel = require('../Models/user.model')
 // const getMyTodolists = async (req, res) => {
 //     try {
 //       const data = await TodolistModel.find({})
@@ -11,46 +11,78 @@ const UserModel = require("../Models/user.model");
 
 const addTodolist = async (req, res) => {
   try {
-    const newList = await TodolistModel.create(req.body.Todolist);
-    const userId = req.payload.id;
+    const userId = req.payload.id
+    if (!(await UserModel.findById(userId))) {
+      return res.json({
+        err: "Invalid Credentials (Id in payload doesn't belong to a user))",
+        statusCode: 1,
+      })
+    }
+    const newList = await TodolistModel.create(req.body.Todolist)
     try {
-      const user = await UserModel.findById(userId);
-      user.lists.push(newList.id);
-      await user.save();
-      return res.json({ msg: "New list added successfully", statusCode: 0 });
+      const user = await UserModel.findById(userId)
+      user.lists.push(newList.id)
+      await user.save()
+      return res.json({ msg: 'New list added successfully', statusCode: 0 })
     } catch (err) {
-      return res.json({ err: "Server Error", statusCode: 1 });
+      return res.json({ err: 'Server Error', statusCode: 1 })
     }
   } catch (err) {
-    return res.json({ err: "Unauthorised User", statusCode: 1 });
+    return res.json({ err: 'Unauthorised User', statusCode: 1 })
   }
-};
+}
 
 const updateTodolist = async (req, res) => {
   try {
+    if (!(await TodolistModel.findById(req.body.Todolist.id))) {
+      return res.json({ err: "id doesn't belong to a list", statusCode: 1 })
+    }
     await TodolistModel.findByIdAndUpdate(req.body.Todolist.id, {
       title: req.body.Todolist.title,
-    });
-    return res.json({ msg: "List updated successfully", statusCode: 0 });
+    })
+    return res.json({ msg: 'List updated successfully', statusCode: 0 })
   } catch (err) {
-    return res.json({ err: "server error", statusCode: 1 });
+    return res.json({ err: 'Server Error', statusCode: 1 })
   }
-};
+}
 
+// PROBLEM (case of wrong id)
 const deleteTodolist = async (req, res) => {
   try {
-    await TodolistModel.findByIdAndDelete(req.body.id);
+    const userId = req.payload.id
+    if (!(await UserModel.findById(userId))) {
+      return res.json({
+        err: "Invalid Credentials (Id in payload doesn't belong to a user))",
+        statusCode: 1,
+      })
+    }
+    // try {
+    //   const list = await TodolistModel.findByIdAndDelete(req.body.id)
+    // } catch (err) {
+    //   return res.json({
+    //     err: 'Resource Not Found',
+    //     statusCode: 1,
+    //   })
+    // }
+    if (!(await TodolistModel.findById(req.body.id))) {
+      return res.json({
+        err: 'Resource Not Found',
+        statusCode: 1,
+      })
+    }
+    await TodolistModel.findByIdAndDelete(req.body.id)
+
     // const user = await UserModel.findById(userId)
     // user.lists.filter((id) => id !== req.body.id)
     // await user.save()
-    const userId = req.payload.id;
+
     await UserModel.findByIdAndUpdate(userId, {
       $pull: { lists: { $in: [req.body.id] } },
-    });
-    return res.json({ msg: "List deleted Successfully", statusCode: 0 });
+    })
+    return res.json({ msg: 'List deleted Successfully', statusCode: 0 })
   } catch (err) {
-    return res.json({ err: "server error", statusCode: 1 });
+    return res.json({ err: 'Server Error', statusCode: 1 })
   }
-};
+}
 
-module.exports = { addTodolist, updateTodolist, deleteTodolist };
+module.exports = { addTodolist, updateTodolist, deleteTodolist }
